Move request payload building to transport adapter
