+++ dh/transportadapter/transportadapter.go
@@ -13,4 +13,5 @@ func New(tsp transport.Transporter) TransportAdapter {
 type TransportAdapter interface {
 	HandleResponseError(rawRes []byte) error
 	ResolveResource(resName string, data map[string]interface{}) (resource, method string)
+	BuildRequestData(resourceName string, rawData map[string]interface{}) interface{}
 }
